refactor: move base layer classes to base_layer.py
The abstract classes LipschitzLayer and Condensable are moved to file
base_layer.py
