+++ deel/lip/layers.py
@@ -49,114 +49,6 @@ except ModuleNotFoundError:
     from tensorflow.python.keras.utils import conv_utils  # in TF.python for TF <= 2.5
 
 
-class LipschitzLayer(abc.ABC):
-    """
-    This class allow to set lipschitz factor of a layer. Lipschitz layer must inherit
-    this class to allow user to set the lipschitz factor.
-
-    Warning:
-         This class only regroup useful functions when developing new Lipschitz layers.
-         But it does not ensure any property about the layer. This means that
-         inheriting from this class won't ensure anything about the lipschitz constant.
-    """
-
-    k_coef_lip = 1.0
-    """variable used to store the lipschitz factor"""
-    coef_lip = None
-    """
-    define correction coefficient (ie. Lipschitz bound ) of the layer
-    ( multiply the output of the layer by this constant )
-    """
-
-    def set_klip_factor(self, klip_factor):
-        """
-        Allow to set the lipschitz factor of a layer.
-
-        Args:
-            klip_factor: the Lipschitz factor the user want to ensure.
-
-        Returns:
-            None
-
-        """
-        self.k_coef_lip = klip_factor
-
-    @abc.abstractmethod
-    def _compute_lip_coef(self, input_shape=None):
-        """
-        Some layers (like convolution) cannot ensure a strict lipschitz constant (as
-        the Lipschitz factor depends on the input data). Those layers then rely on the
-        computation of a bounding factor. This function allow to compute this factor.
-
-        Args:
-            input_shape: the shape of the input of the layer.
-
-        Returns:
-            the bounding factor.
-
-        """
-        pass
-
-    def _init_lip_coef(self, input_shape):
-        """
-        Initialize the lipschitz coefficient of a layer.
-
-        Args:
-            input_shape: the layers input shape
-
-        Returns:
-            None
-
-        """
-        self.coef_lip = self._compute_lip_coef(input_shape)
-
-    def _get_coef(self):
-        """
-        Returns:
-            the multiplicative coefficient to be used on the result in order to ensure
-            k-Lipschitzity.
-        """
-        if self.coef_lip is None:
-            raise RuntimeError("compute_coef must be called before calling get_coef")
-        return self.coef_lip * self.k_coef_lip
-
-
-class Condensable(abc.ABC):
-    """
-    Some Layers don't optimize directly the kernel, this means that the kernel stored
-    in the layer is not the kernel used to make predictions (called W_bar), to address
-    this, these layers can implement the condense() function that make self.kernel equal
-    to W_bar.
-
-    This operation also allow the turn the lipschitz layer to it keras equivalent ie.
-    The Dense layer that have the same predictions as the trained SpectralDense.
-    """
-
-    @abc.abstractmethod
-    def condense(self):
-        """
-        The condense operation allow to overwrite the kernel and ensure that other
-        variables are still consistent.
-
-        Returns:
-            None
-
-        """
-        pass
-
-    @abc.abstractmethod
-    def vanilla_export(self):
-        """
-        This operation allow to turn this Layer to it's super type, easing storage and
-        serving.
-
-        Returns:
-             self as super type
-
-        """
-        pass
-
-
 def _check_RKO_params(eps_spectral, eps_bjorck, beta_bjorck):
     """Assert that the RKO hyper-parameters are supported values."""
     if eps_spectral <= 0:
